#137 Sort out presentation vs. publication helpers, add some specs, support publication in menu
