+++ app/helpers/presentations_helper.rb
@@ -1,45 +1,7 @@
 module PresentationsHelper
 
-  def format_and_date(publication)
-    if publication.url.present?
-      link_to(icon_for_format(publication)  + ' ' + publication.format, publication.url, target: '_blank') + ", #{ publication.published_on.year }"
-    else
-      "#{ icon_for_format(publication) } #{ publication.format }, #{ publication.published_on.year }".html_safe
-    end
-  end
-
   def linked_tag_names(presentation)
     presentation.tag_list.map{|t| link_to t, presentations_path(tag: t)}.join(', ').html_safe
-  end
-
-  def icon_for_format(publication)
-    icon = case publication.format
-    # Put the tooltip directly on the icons that are unlikely to have links - TODO - enforce which does and doesn't get a link
-    when Publication::TAPE then
-      icon('fas', 'tape', 'data-toggle' => "tooltip", title: publication.notes)
-    when Publication::CD then
-      icon('fas', 'compact-disc', 'data-toggle' => "tooltip", title: publication.notes)
-    when Publication::VHS then
-      icon('fas', 'ticket-alt', 'data-toggle' => "tooltip", title: publication.notes)
-    when Publication::DISK then
-      icon('fas', 'compact-disc', 'data-toggle' => "tooltip", title: publication.notes)
-    when Publication::CAMPUS then
-      icon('fas', 'university')
-    when Publication::YOUTUBE then
-      icon('fab', 'youtube')
-    when Publication::FACEBOOK then
-      icon('fab', 'facebook')
-    when Publication::PODCAST then
-      icon('fas', 'podcast')
-    when Publication::ONLINE then
-      icon('fas', 'download')
-    when Publication::ESTORE then
-      icon('fas', 'store')
-    when Publication::PRINT then
-      icon('fas', 'book')
-    else
-      icon('fas', 'question-circle')  # This means something was added to Publication FORMATS but not included here
-    end
   end
 
   # Get a list of icons corresponding to items in the publication list. The ones that correspond to online media are
@@ -47,9 +9,7 @@ module PresentationsHelper
   def linked_format_icons(presentation)
     icon_list = []
     presentation.publications.each do |publication|
-      icon = icon_for_format(publication)
-      link_text = publication.url.present? ? link_to(icon, publication.url, 'data-toggle' => "tooltip", title: publication.notes, target: '_blank') : icon # individual publications may or may not be links
-      icon_list << link_text
+      icon_list << linked_icon(publication)
     end
 
     return icon_list.join(' ').html_safe
