+++ spec/helpers/presentations_helper_spec.rb
@@ -1,15 +1,11 @@
 require 'rails_helper'
 
-# Specs in this file have access to a helper object that includes
-# the PresentationsHelper. For example:
-#
-# describe PresentationsHelper do
-#   describe "string concat" do
-#     it "concats two strings with spaces" do
-#       expect(helper.concat_strings("this","that")).to eq("this that")
-#     end
-#   end
-# end
 RSpec.describe PresentationsHelper, type: :helper do
-  pending "add some examples to (or delete) #{__FILE__}"
+
+  let(:presentation) { create :presentation }
+
+  describe "generating a list of linked speaker names" do
+    it "generates the list"
+  end
+
 end
